Add rentals manager and seperate classes into files
