+++ app.rb
@@ -2,16 +2,18 @@ require './book'
 require './rental'
 require './student'
 require './teacher'
+require './books_manager'
+require './rentals_manager'
 require 'date'
 
 class App
   def initialize()
-    @books_manager = BooksManager.new([])
-    @people = []
-    @rentals = []
+    @books_manager = BooksManager.new([Book.new('The call of Cthulu', 'H.P. Lovecraft')])
+    @people = [Student.new(12, 'Ash')]
+    @rentals_manager = RentalsManager.new([], @books_manager, @people)
   end
 
-  def useMethod(user_input)
+  def use_method(user_input)
     case user_input
     when 1
       @books_manager.list_books
@@ -22,12 +24,12 @@ class App
     when 4
       @books_manager.create_book
     when 5
-      create_rental
+      @rentals_manager.create_rental
     when 6
       puts 'Please enter the id of the person'
       print 'Id: '
       id = gets.chomp
-      list_rentals_of(id.to_i)
+      @rentals_manager.list_rentals_of(id.to_i)
     else
       puts "How'd you do that?"
     end
@@ -72,73 +74,5 @@ class App
     specialization = gets.chomp.capitalize
     @people.push(Teacher.new(specialization, age, name))
     puts 'The teacher was added successfully'
-  end
-
-  def create_rental
-    if @books.empty?
-      puts 'There are no books in the library'
-      return
-    end
-
-    if @people.empty?
-      puts 'There are no people in the library'
-      return
-    end
-
-    book = select_book
-    person = select_person
-
-    current_date = Date.today
-    @rentals.push(Rental.new(current_date, book, person))
-    puts current_date
-    puts 'The rental instance was created successfully'
-  end
-
-  def select_book
-    puts 'Select a book from the following list by number'
-    @books.each_with_index { |book, index| puts " #{index}) Title: #{book.title}, Author: #{book.author}" }
-    book_number = gets.chomp.to_i
-    @books[book_number]
-  end
-
-  def select_person
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index do |person, index|
-      tag = person.is_a?(Student) ? 'Student' : 'Teacher'
-      puts " #{index}) [#{tag}] ID: #{person.id} Name: #{person.name}, Age: #{person.age}"
-    end
-    person_number = gets.chomp.to_i
-    @people[person_number]
-  end
-
-  def list_rentals_of(id)
-    puts 'Rentals:'
-    rentals = @rentals.select { |rental| rental.person.id == id }
-    if rentals.any?
-      rentals.map { |rental| puts "Date: #{rental.date}, Book: '#{rental.book.title}' by '#{rental.book.author}'" }
-    else
-      puts 'No rentals found for the specified person ID.'
-    end
-  end
-end
-
-class BooksManager
-  def initialize(books)
-    @books = books
-  end
-
-  def list_books
-    @books.map { |book| puts "Title: #{book.title}, Author: #{book.author}" }
-  end
-  
-  def create_book
-    puts "Please enter the book's"
-    print 'Title: '
-    title = gets.chomp
-    print 'Author: '
-    author = gets.chomp
-    new_book = Book.new(title, author)
-    @books.push(new_book)
-    print "The book '#{new_book.title}' by '#{new_book.author}' was added successfully"
   end
 end
