+++ books_manager.rb
@@ -1,0 +1,22 @@
+class BooksManager
+  attr_reader :books
+
+  def initialize(books)
+    @books = books
+  end
+
+  def list_books
+    @books.map { |book| puts "Title: '#{book.title}', Author: '#{book.author}'" }
+  end
+
+  def create_book
+    puts "Please enter the book's"
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    new_book = Book.new(title, author)
+    @books.push(new_book)
+    print "The book '#{new_book.title}' by '#{new_book.author}' was added successfully"
+  end
+end
